+++ cmd/awooll/token/variants.go
@@ -1,41 +1,41 @@
 package token
 
 // General.
-const TokenTypeIdentifier = 0x000
-const TokenTypePrimitive = 0x001
-const TokenTypeType = 0x002
-const TokenTypeEndStatement = 0x003
-const TokenTypeBracketLeft = 0x004
-const TokenTypeBracketRight = 0x005
-const TokenTypeBracketCurlyLeft = 0x006
-const TokenTypeBracketCurlyRight = 0x007
-const TokenTypeNot = 0x008
-const TokenTypeComma = 0x009
-const TokenTypeBracketSquareLeft = 0x00A
-const TokenTypeBracketSquareRight = 0x00B
+const TokenTypeIdentifier = uint16(0x000)
+const TokenTypePrimitive = uint16(0x001)
+const TokenTypeType = uint16(0x002)
+const TokenTypeEndStatement = uint16(0x003)
+const TokenTypeBracketLeft = uint16(0x004)
+const TokenTypeBracketRight = uint16(0x005)
+const TokenTypeBracketCurlyLeft = uint16(0x006)
+const TokenTypeBracketCurlyRight = uint16(0x007)
+const TokenTypeNot = uint16(0x008)
+const TokenTypeComma = uint16(0x009)
+const TokenTypeBracketSquareLeft = uint16(0x00A)
+const TokenTypeBracketSquareRight = uint16(0x00B)
 
 func IsTokenTypeGeneral(t uint16) bool {
 	return t < 0x100
 }
 
 // Operators.
-const TokenOperatorAddition = 0x100
-const TokenOperatorSubstraction = 0x101
-const TokenOperatorMultiplication = 0x102
-const TokenOperatorDereference = 0x102
-const TokenOperatorDivision = 0x103
-const TokenOperatorEq = 0x104
-const TokenOperatorEqEq = 0x105
-const TokenOperatorNotEq = 0x106
-const TokenOperatorLT = 0x107
-const TokenOperatorLTEQ = 0x108
-const TokenOperatorGT = 0x109
-const TokenOperatorGTEQ = 0x110
-const TokenOperatorLS = 0x111
-const TokenOperatorRS = 0x112
-const TokenOperatorAnd = 0x113
-const TokenOperatorReference = 0x113
-const TokenOperatorOr = 0x114
+const TokenOperatorAddition = uint16(0x100)
+const TokenOperatorSubstraction = uint16(0x101)
+const TokenOperatorMultiplication = uint16(0x102)
+const TokenOperatorDereference = uint16(0x102)
+const TokenOperatorDivision = uint16(0x103)
+const TokenOperatorEq = uint16(0x104)
+const TokenOperatorEqEq = uint16(0x105)
+const TokenOperatorNotEq = uint16(0x106)
+const TokenOperatorLT = uint16(0x107)
+const TokenOperatorLTEQ = uint16(0x108)
+const TokenOperatorGT = uint16(0x109)
+const TokenOperatorGTEQ = uint16(0x110)
+const TokenOperatorLS = uint16(0x111)
+const TokenOperatorRS = uint16(0x112)
+const TokenOperatorAnd = uint16(0x113)
+const TokenOperatorReference = uint16(0x113)
+const TokenOperatorOr = uint16(0x114)
 
 func IsTokenTypeOperator(t uint16) bool {
 	return t >= 0x100 && t < 0x200
@@ -66,13 +66,13 @@ func DoesTokenTakePrecendence(op uint16, left uint16) bool {
 }
 
 // Keywords.
-const TokenTypeVar = 0x200
-const TokenTypeTypeDefinition = 0x201
-const TokenTypeIf = 0x202
-const TokenTypeElse = 0x203
-const TokenTypeFunc = 0x204
-const TokenTypeReturn = 0x205
-const TokenTypeFor = 0x206
+const TokenTypeVar = uint16(0x200)
+const TokenTypeTypeDefinition = uint16(0x201)
+const TokenTypeIf = uint16(0x202)
+const TokenTypeElse = uint16(0x203)
+const TokenTypeFunc = uint16(0x204)
+const TokenTypeReturn = uint16(0x205)
+const TokenTypeFor = uint16(0x206)
 
 func IsTokenTypeKeyword(t uint16) bool {
 	return t >= 0x200 && t < 0x300
