+++ cmd/common/instruction/load_store.go
@@ -1,12 +1,12 @@
 package instruction
 
-var AwooInstructionsLoad = map[uint16]*AwooInstruction{
+var AwooInstructionsLoad = map[uint32]*AwooInstruction{
 	1: &AwooInstructionLB,
 	2: &AwooInstructionLH,
 	4: &AwooInstructionLW,
 }
 
-var AwooInstructionsSave = map[uint16]*AwooInstruction{
+var AwooInstructionsSave = map[uint32]*AwooInstruction{
 	1: &AwooInstructionSB,
 	2: &AwooInstructionSH,
 	4: &AwooInstructionSW,
